penambahan icon shoppping cart

+++ database/migrations/2024_06_17_135659_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->text('description');
             $table->string('image');
-            $table->decimal('price',10,0);
+            $table->integer('price');
             $table->timestamps();
         });
     }
